moved course query to the shared library
