+++ clients/shared_library/interfaces/course.ts
